+++ internal/domain/armorXCharacterData.go
@@ -1,8 +1,8 @@
 package domain
 
 type ArmorXCharacterData struct {
-	ArmorXCharacterData_Id int64 `json:"armorxcharacterdataid"`
+	ArmorXCharacterData_Id int   `json:"armorxcharacterdataid"`
 	Armor_Id               Armor `json:"armor"`
-	CharacterData_Id       int64 `json:"characterdataid"`
+	CharacterData_Id       int   `json:"characterdataid"`
 	Equipped               bool  `json:"equipped"`
 }
